[1-9]: Part1: handle content-encoding

+++ response.go
@@ -9,7 +9,7 @@ import (
 type response struct {
 	proto        string
 	conn         io.ReadWriteCloser
-	status       int64
+	status       int
 	statusString string
 	headers      map[string]string
 	content      string
